sidecar/python/requirements.txt
# Presence Layer Dependencies
livekit-agents[silero,turn-detector]~=1.4
livekit-plugins-ollama
livekit-plugins-silero
kokoro-onnx
soundfile
python-dotenv
aiohttp
